relocate files and create empty signin page
