+++ frontend/lib/models/truck_model.dart
