+++ frontend/lib/views/truck_map_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'dart:async';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'truck_model.dart';
+import '../models/truck_model.dart';
 
 class FoodTruckMapView extends StatefulWidget {
