+++ frontend/lib/widgets/display_type_tab.dart
@@ -7,7 +7,7 @@
 
 
 import 'package:flutter/material.dart';
-import 'utils/Utils.dart';
+import '../utils/Utils.dart';
 
 
 class ExploreViewTypeTab extends StatelessWidget {
